requirements.txt
python-dateutil>=1.5,<2.5
requests-oauthlib>=0.4,<1.1